requirements.txt
gradio>=3.50
crewai>=0.1.0
langchain>=0.1.0
langchain-community>=0.1.0
google-generativeai>=0.2.0
python-dotenv>=1.0.0
